向量调整为64位

+++ src/Face.py
@@ -28,7 +28,7 @@ class Face():  #基类，包含人脸编码，人脸识别
 class StudentRgFace(Face):
     def __init__(self):
         super().__init__()
-        self.face_data = np.random.random(128).astype('float32')#初始化人脸编码，这个变量保存上一个人脸编码
+        self.face_data = np.random.random(128).astype('float64')#初始化人脸编码，这个变量保存上一个人脸编码
         self.former_result = ""
         self.refreshthread = Timer(3, self.reset)
         self.refreshthread.setDaemon(True)
@@ -41,7 +41,7 @@ class StudentRgFace(Face):
       
         student.conn.close()
     def reset(self):
-        self.face_data = np.random.random(128).astype('float32')
+        self.face_data = np.random.random(128).astype('float64')
         self.refreshthread = Timer(3, self.reset)
         self.refreshthread.setDaemon(True)
         self.refreshthread.start()
